+++ view/frontend/web/js/view/payment/hipay-methods.js
@@ -93,6 +93,11 @@ define([
         'HiPay_FullserviceMagento/js/view/payment/method-renderer/hipay-ing'
     },
     {
+      type: 'hipay_ideal_hosted_fields',
+      component:
+        'HiPay_FullserviceMagento/js/view/payment/method-renderer/hipay-ideal-hosted-fields'
+    },
+    {
       type: 'hipay_ideal',
       component:
         'HiPay_FullserviceMagento/js/view/payment/method-renderer/hipay-ideal'